Refactor project directory structure
