--- extension/manifest.json
@@ -1,16 +1,0 @@
-{
-  "name": "Git Achievements",
-  "description": "Rewards the user with achievements upon completions of challenges.",
-  "version": "1.0.0",
-  "manifest_version": 2,
-  "homepage_url": "https://github.com/KennyyNguyen/Git-Achievements",
-  "browser_action": {
-    "default_title": "Git Achievements",
-    "default_popup": "index.html"
-  },
-  "icons": {
-    "16": "../public/img/icon-16.png",
-    "48": "../public/img/icon-48.png",
-    "128": "../public/img/icon-128.png"
-  }
-}
