+++ src/common/initGitlabApi.js
