+++ src/common/updateSetting.js
@@ -1,6 +1,6 @@
 import * as browser from "webextension-polyfill";
 
-export default async function updateSetting(objectToStore) {
+export const updateSetting = async (objectToStore) => {
   await browser.storage.local.set(objectToStore);
-  console.log("Settings updated!");
-}
+  console.log("Setting updated!");
+};
